+++ app/presentation/auth/login/login_endpoints.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter, Depends, Response, status
 
 from app.domain.usecase.auth.login_usecase import LoginUseCase
-from app.presentation.login.login_request import LoginRequest
-from app.presentation.login.login_response import LoginResponse
-from app.presentation.login.login_translator import LoginTranslator
+from app.presentation.auth.login.login_request import LoginRequest
+from app.presentation.auth.login.login_response import LoginResponse
+from app.presentation.auth.login.login_translator import LoginTranslator
 from app.di.usecase import inject_login_usecase
 
 router = APIRouter()
