+++ main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 
-from app.presentation.login import login_endpoints
+from app.presentation.auth.login import login_endpoints
 from app.presentation.user import user_endpoints
